Trunk: Merged rtma_nlqc2 at r64212 to the trunk. In the ticket #291, Xiujuan Su documented the task and code development for adapting Jim Purser's non-linear quality control scheme (denoted as "New VarQC" in this ticket). This ticket, with branch name: rtma_nlqc2, describes the work of applying Su's latest repository branch (Feb. 2015) into Real-Time Mesoscale Analysis (RTMA) system. The major differences in this application are (1) RTMA analysis is surface 2D analysis;(2) dominant observation type is MESONET, which is more than 80% of the total observation types. We started to tune parameters for MESONET.
One more change into this coming upgrade RTMA (developed by Pondeca and Levine): the current code assumes that the minimum and maximum temperature observations are all read in from the prepbufr file. A change is made to add the ability to read such observations from an ascii file as well. Subroutine "read_mitm_mxtm.f90" is added to that effect.

For further details, please see ticket #493.


git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/src@64534 281fb692-3a18-0410-826d-9e847983f365