Trunk: Merged fixe2c branch at r64861 to the trunk.  During the course of testing it was found that the pr4devb was slightly lower quality than the pr4dev. One of the three differences in addition to the land surface changes and a zeroing out of negative clw retrievals is the "fixing" of an inconsistency in the ensctl2state and ensctl2model routines. Closer examination shows that this fix is not correct and does not interact with the static term properly.
It ia uncertain whether the original is completely correct, but it is better than the "fix" that was put in, since the fix included an error in the interaction of the static term and the ensemble term, while the original had just a possible small inconsistency between the moisture and sensible temperature variables.

The new branch, fixe2c, was created to remove the modifications. Note some of the changes to these routines were good so not all modifications were removed.
Fixed bug in normal_rh_to_q. Then modified check for qoption = 2 to be qoption == 2 rather than qoption /= 1.

Included a fix to a bug in setupbend.f90 found by Russ. 

For further details, please see ticket #517.



git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/src@65438 281fb692-3a18-0410-826d-9e847983f365