Trunk: Merged minim branch at r67522 to the trunk.  The global pre-implementation parallel has been showing occasional issues in the minimization. Many resets and occasional early stopping. Looking at the output indicates that much of the issue occurs because of competition between the supersaturation constraint and the other moisture terms. The minimization would work properly when the supersaturation constraint was turned off.
It does not appear that there is any problem with the supersaturation constraint, but rather the issue is the background error covariance for the q field. By putting a simple limit on the input moisture (0<=q<=1.2*qsat) in the ensembles, the problem went away. However this limit has some problems, because if all of the ensembles are greater than supersaturation, or less than zero, the variance becomes zero and no change is allowed in the analysis at this location. 

A second solution was tried. In this, the difference of the mean ensemble field at each analysis time (e.g., -3,-2,-1,0,1,2,3hr) was limited to be between -100% and 100%. This resulted in a clean minimization and much improved fits to obs for all variables for the cases the minimization failed.

Russ ran this version on other cases where there were not minimization issues. The results did change, but no significant differences were noted.

Those points that were adjusted were printed out. They showed values that input values that were often > 200% of saturation (sometimes much greater). This seems very unphysical and may mean we need to revisit how we are perturbing the ensembles and/or when we are writing the forecasts. Currently we are writing the output immediately after the perturbation. So if the q field is perturbed to greater than saturation, the physics does not have a chance to precipitate the additional moisture.

Changes necessary to limit the difference from the mean to +-100% RH are put into branch minim.

Additionally, the aircraft bias correction could play a role in the "Resets" occurrences. 

After reviewing the GSI routines, it was discovered that the preconditioning for the aircraft bias correction was only applied to the cases when obs_count>3, but not to the cases when obs_count<=3 as it was undesirable to use rstats_t information from these very few observations in the preconditioning formula when the code was originally written. It is now suspected that this may cause a problem for an observation from preconditioned in a cycle to un-preconditioned to the next cycle, so it was suggested to apply the preconditioning formula to the cases with obs_count<3 as well.

Only two routines were modified in this update:

 * src:
   * get_gefs_ensperts_dualres.f90
   * berror.f90


git-svn-id: https://svnemc.ncep.noaa.gov/projects/gsi/trunk/src@67694 281fb692-3a18-0410-826d-9e847983f365